feat: move timer cb
