+++ driving_log_replayer/scripts/eagleye_evaluator_node.py
@@ -14,23 +14,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-from typing import TYPE_CHECKING
-
 from diagnostic_msgs.msg import DiagnosticArray
 from diagnostic_msgs.msg import DiagnosticStatus
 import rclpy
 from rclpy.executors import MultiThreadedExecutor
-from rclpy.node import Node
-from tier4_localization_msgs.srv import PoseWithCovarianceStamped
 import yaml
 
 from driving_log_replayer.evaluator import DLREvaluator
 from driving_log_replayer.result import ResultBase
 from driving_log_replayer.result import ResultWriter
-
-if TYPE_CHECKING:
-    from autoware_adapi_v1_msgs.msg import ResponseStatus
 
 
 class EagleyeResult(ResultBase):
@@ -86,7 +78,7 @@ class EagleyeEvaluator(DLREvaluator):
         self._initial_pose = DLREvaluator.set_initial_pose(
             self._scenario_yaml_obj["Evaluation"]["InitialPose"]
         )
-        self.start_initial_pose()
+        self.start_initialpose_service()
 
         self.__sub_diagnostics = self.create_subscription(
             DiagnosticArray,
@@ -98,33 +90,6 @@ class EagleyeEvaluator(DLREvaluator):
     def diagnostics_cb(self, msg: DiagnosticArray):
         self.__result.add_eagleye_availability_frame(msg)
         self.__result_writer.write(self.__result)
-
-    def timer_cb(self) -> None:
-        self._current_time = self.get_clock().now().to_msg()
-        # self.get_logger().error(f"time: {self.__current_time.sec}.{self.__current_time.nanosec}")
-        if self._current_time.sec > 0:
-            if (
-                self._initial_pose is not None
-                and not self._initial_pose_success
-                and not self._initial_pose_running
-            ):
-                self.get_logger().info(
-                    f"call initial_pose time: {self._current_time.sec}.{self._current_time.nanosec}"
-                )
-                self._initial_pose_running = True
-                self._initial_pose.header.stamp = self._current_time
-                future_map_fit = self._map_fit_client.call_async(
-                    PoseWithCovarianceStamped.Request(pose_with_covariance=self._initial_pose)
-                )
-                future_map_fit.add_done_callback(self.map_fit_cb)
-            if self._current_time == self._prev_time:
-                self._clock_stop_counter += 1
-            else:
-                self._clock_stop_counter = 0
-            self._prev_time = self._current_time
-            if self._clock_stop_counter >= 5:
-                self._result_writer.close()
-                rclpy.shutdown()
 
 
 def main(args=None):
